requirements.txt
glob2
numpy
pillow
scipy